Renamed Pager to Pages
